requirements.txt
yfinance>=0.2.18
pandas>=1.4.0
openpyxl>=3.0.0
python-dateutil>=2.8.2 